v2.0.0 enhancement: Add color for users & sys msgs
Made it so it selects a color from the list. This list is based on the stylesh package, and makes sure the selected color is a string to pass as a variable. Currently we're only coloring the border of the user, but could easily do the name itself, or the background of the text.  Tried to make it as backwards compatible as possible.